+++ app/includes/word_picture.pug
@@ -4,14 +4,14 @@ div.wordpic_disabled(ng-if="!word_pic")
     button.btn.btn-sm.btn-default.activate_word_pic(ng-click="activate()")
       | {{'word_pic_warn_btn' | loc:lang}}
 
-warning(ng-if="word_pic && !instance.hasData && !loading && !aborted")
+warning(ng-if="word_pic && !hasData && !loading && !aborted")
   | {{'word_pic_bad_search' | loc:lang}}
 warning(ng-if="word_pic && aborted && !loading")
   | {{'search_aborted' | loc:lang}}
 warning(ng-if="word_pic && no_hits")
   | {{"no_stats_results" | loc:lang}}
-div(constr='WordPictureResults', constr-name='wordPictureResults')
-  #wordPicSettings(ng-show="word_pic && instance.hasData")
+div
+  #wordPicSettings(ng-show="word_pic && hasData")
     div()
       input#wordclassChk(ng-model="showWordClass", type='checkbox')
       label(for='wordclassChk') {{'show_wordclass' | loc:lang}}
